// src/app/core/models/LSS_models/Documentation.ts
import { ImageDocum } from "./ImageDocum";

export interface Documentation {
    iddocu: string;
    sectiondocu: string;
    contentdocu: string;
    lastModifieddocu: Date;
    content1: string;
    content2: string;
    content3: string;
    content4: string;
    content5: string;
    content6: string;
    content7: string;
    content8: string;
    title1: string;
    title2: string;
    title3: string;
    title4: string;
    title5: string;
    title6: string;
    title7: string;

    imagesdocu: ImageDocum[];

  }